core/src/main/java/com/orientechnologies/orient/core/storage/impl/local/paginated/wal/component/sbtree/page/sbtreebucket/OSBTreeBucketSetRightSiblingOperation.java
package com.orientechnologies.orient.core.storage.impl.local.paginated.wal.component.sbtree.page.sbtreebucket;

import com.orientechnologies.common.serialization.types.OLongSerializer;
import com.orientechnologies.orient.core.storage.impl.local.paginated.wal.OLogSequenceNumber;
import com.orientechnologies.orient.core.storage.impl.local.paginated.wal.OPageOperation;
import com.orientechnologies.orient.core.storage.impl.local.paginated.wal.WALRecordTypes;

import java.nio.ByteBuffer;

public class OSBTreeBucketSetRightSiblingOperation extends OPageOperation {
  private long rightSibling;

  public OSBTreeBucketSetRightSiblingOperation() {
  }

  public OSBTreeBucketSetRightSiblingOperation(OLogSequenceNumber pageLSN, long fileId, long pageIndex, long rightSibling) {
    super(pageLSN, fileId, pageIndex);
    this.rightSibling = rightSibling;
  }

  @Override
  public byte getId() {
    return WALRecordTypes.SBTREE_BUCKET_SET_RIGHT_SIBLING_OPERATION;
  }

  @Override
  public int serializedSize() {
    return super.serializedSize() + OLongSerializer.LONG_SIZE;
  }

  @Override
  public int toStream(byte[] content, int offset) {
    offset = super.toStream(content, offset);

    OLongSerializer.INSTANCE.serializeNative(rightSibling, content, offset);
    offset += OLongSerializer.LONG_SIZE;

    return offset;
  }

  @Override
  public void toStream(ByteBuffer buffer) {
    super.toStream(buffer);

    buffer.putLong(rightSibling);
  }

  @Override
  public int fromStream(byte[] content, int offset) {
    offset = super.fromStream(content, offset);

    rightSibling = OLongSerializer.INSTANCE.deserializeNative(content, offset);
    offset += OLongSerializer.LONG_SIZE;

    return offset;
  }

  public long getRightSibling() {
    return rightSibling;
  }
}
